// src/app/core/services/models/dashboard-stats.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface DashboardStats {
  cancelledMissions?: number;
  completedMissions?: number;
  inProgressMissions?: number;
  missionsByMonth?: {
[key: string]: number;
};
  missionsBySkill?: {
[key: string]: number;
};
  missionsByUser?: {
[key: string]: number;
};
  pendingMissions?: number;
  totalMissions?: number;
  totalUsers?: number;
}